# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads/session/{token}:
    get:
      operationId: getSessionByToken
      tags: [leads]
      summary: Get conversation by session token
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /leads/{id}/complete:
    post:
      operationId: completeLead
      tags: [leads]
      summary: Mark lead as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadCompleteResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/leads:
    get:
      operationId: listAdminLeads
      tags: [admin]
      summary: List all leads
      parameters:
        - name: personaId
          in: query
          required: false
          schema:
            type: integer
        - name: personaTypeId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeadSummary"
  /admin/leads/{id}:
    get:
      operationId: getAdminLead
      tags: [admin]
      summary: Get lead detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteAdminLead
      tags: [admin]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /leads/{id}/contact:
    post:
      operationId: saveLeadContact
      tags: [leads]
      summary: Save contact details supplied by the visitor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInfoInput"
      responses:
        "200":
          description: Contact saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSaveResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /leads/{id}/attachments:
    get:
      operationId: listLeadAttachments
      tags: [leads]
      summary: List files attached to a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Attachments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attachment"
...
  /leads/upload:
    post:
      operationId: uploadLeadAttachment
      tags: [leads]
      summary: Attach a file to a conversation
      requestBody:
        required: true
        content:
          multipart/form-data: {}
      responses:
        "201":
          description: File stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttachmentUploadResult"
        "400":
          description: Missing or rejected file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
    AttachmentUploadResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        id:
          type: integer
        fileName:
          type: string
        fileUrl:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
      required: [success, message, id, fileName, fileUrl, mimeType, sizeBytes]
...
    ContactInfo:
      type: object
      properties:
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
    ContactInfoInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: string
    ContactSaveResult:
      type: object
      properties:
        success:
          type: boolean
        contactInfo:
          $ref: "#/components/schemas/ContactInfo"
      required: [success, contactInfo]
    Attachment:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: ["integer", "null"]
        fileName:
          type: string
        fileUrl:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, fileName, fileUrl, mimeType, sizeBytes, createdAt]
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    AnthropicMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, conversationId, role, content, createdAt]
    AnthropicConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        sessionToken:
          type: string
        emailSent:
          type: boolean
        completed:
          type: boolean
        summary:
          type: ["string", "null"]
        personaId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AnthropicMessage"
      required: [id, title, sessionToken, emailSent, completed, createdAt, messages]
...
    LeadCompleteResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        emailSent:
          type: boolean
        webhookSent:
          type: boolean
        summary:
          type: ["string", "null"]
        contactInfo:
          $ref: "#/components/schemas/ContactInfo"
      required: [success, message]
    LeadSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        sessionToken:
          type: string
        completed:
          type: boolean
        summary:
          type: ["string", "null"]
        requirements:
          type: ["string", "null"]
        contactInfo:
          $ref: "#/components/schemas/ContactInfo"
        emailSent:
          type: boolean
        webhookSent:
          type: boolean
        messageCount:
          type: integer
        personaId:
          type: ["integer", "null"]
        personaName:
          type: ["string", "null"]
        personaTypeName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, title, sessionToken, completed, contactInfo, messageCount, createdAt]
    LeadDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        sessionToken:
          type: string
        completed:
          type: boolean
        summary:
          type: ["string", "null"]
        requirements:
          type: ["string", "null"]
        contactInfo:
          $ref: "#/components/schemas/ContactInfo"
        emailSent:
          type: boolean
        webhookSent:
          type: boolean
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
        messageCount:
          type: integer
        personaId:
          type: ["integer", "null"]
        personaName:
          type: ["string", "null"]
        personaTypeName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AnthropicMessage"
      required: [id, title, sessionToken, completed, contactInfo, messageCount, createdAt, messages]
...