# lib/api-zod/src/generated/types/attachmentUploadResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AttachmentUploadResult {
  success: boolean;
  message: string;
  id: number;
  fileName: string;
  fileUrl: string;
  mimeType: string;
  sizeBytes: number;
}