# lib/api-zod/src/generated/types/leadCompleteResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContactInfo } from './contactInfo';

export interface LeadCompleteResult {
  success: boolean;
  message: string;
  emailSent?: boolean;
  webhookSent?: boolean;
  /** @nullable */
  summary?: string | null;
  contactInfo?: ContactInfo;
}