# lib/api-zod/src/generated/types/leadSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContactInfo } from './contactInfo';

export interface LeadSummary {
  id: number;
  title: string;
  sessionToken: string;
  completed: boolean;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  requirements?: string | null;
  contactInfo: ContactInfo;
  emailSent?: boolean;
  webhookSent?: boolean;
  messageCount: number;
  /** @nullable */
  personaId?: number | null;
  /** @nullable */
  personaName?: string | null;
  /** @nullable */
  personaTypeName?: string | null;
  createdAt: Date;
}